+++ R/liana_wrapper.R
@@ -16,10 +16,9 @@
 #'
 #' @param return_all whether to return all possible interactions. Any
 #' interaction with `expr_prop` below the specific threshold will be
-#' assigned to the *worst* possible score in the results. For example, p-values
-#' from CellPhoneDB will be assigned to max(pvalue) - likely 1, and lr_means
-#' will be assigned to min(lr_means). Note that this might results in slower
-#' computation speed, as all interactions scores need to be first calculated.
+#' assigned to the *worst* possible score in those that pass the threshold.
+#'  For example, p-values from CellPhoneDB will be assigned to max(pvalue)
+#'   - likely 1, and lr_means will be assigned to min(lr_means).
 #'
 #' @param supp_columns any supplementary/additional columns which are to be
 #' returned by liana. Possibilities include: c("ligand.expr", "receptor.expr"
@@ -393,7 +392,9 @@ show_resources <- function(){
         anti_join(non_expressed, by=c("ligand.complex", "receptor.complex",
                                       "source", "target"))
     } else if(return_all){
-      liana_res
+      liana_res %>%
+        mutate(to.filter = receptor.prop >= expr_prop &
+                 ligand.prop >= expr_prop)
     }
 
   } else if(expr_prop == 0){